requirements.txt
# Core
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
aiofiles

# Document Processing
pdfplumber
camelot-py[cv]
tabula-py
pytesseract
Pillow
pandas
openpyxl

# AI / LLM
langchain
langchain-huggingface
langchain-community

# Machine Learning
scikit-learn
xgboost
shap
joblib

# Network Analysis
networkx

# CAM Report Generation
python-docx
reportlab
matplotlib

# Research APIs
requests
gnews

# Web Search (fallback)
duckduckgo-search

# Environment
python-dotenv